Rename 'spell_language' to 'spelling_language'
Discussion on the editorconfig standard seems to converge
on calling the property 'spelling_language' rather than
'spell_language' (see [1]). This commit adjusts both the
value that is expected to be found in .editorconfig files
as well as the internal variable that stores this property.

Note to editorconfig users: This commit changes the plugin's
behaviour in a non backwards compatible way. All editorconfig
files that contain the 'spell_language' property need to be
updated.

Note to spellchecker plugin developers: Since this commit
also updates the internal variable name, plugins need to
update this name as well.

[1] https://github.com/editorconfig/specification/pull/41