Remove FILE_SAVE_POST hook
This hook was intended to be used to re-parse editorconfig
files after writing changes to disk. The idea was that, if the
file being written to disk should be an editorconfig file,
the changes to that file might be applied directly. Since
this does not seem to work properly at the moment, removing
this hook should at least make the code more straightforward.